Non-graphical, includes sqlite database
This is a rewrite of the main program, removing all the tkinter graphical elements (run this as python3 lcc_simple.py and type in the filename when prompted and also creates/uses a local sqlite database to store all the ones you previously found, and uses that to avoid looking up the same title twice. I can't share that file here because it is too large for github, but contact me to get a copy of UPEI's retrieved LCCs.
To run this in your local python, be sure you've added all of the modules listed at the top, including isbnlib and requests which do not come with python.